Build MultiTag from the included source files. Update MultiTag.

+++ burn/Frameworks/Sources/MultiTag/libid3/id3V2Frame.h
@@ -51,9 +51,9 @@
 
 -(NSData *)getRawFrameData;
 -(NSData *)getCompleteRawFrame;
--(NSData *)write2FrameLength:(int)Length;
--(NSData *)write3FrameLength:(int)Length;
--(NSData *)write4FrameLength:(int)Length;
+-(NSData *)write2FrameLength:(NSInteger)Length;
+-(NSData *)write3FrameLength:(NSInteger)Length;
+-(NSData *)write4FrameLength:(NSInteger)Length;
 
 -(BOOL)compress;
 -(BOOL)unsynch;
